DBZ-3086 Remove parallel snapshot reader offsets
