+++ debezium-connector-mysql/src/test/java/io/debezium/connector/mysql/SourceInfoTest.java
@@ -6,7 +6,6 @@
 package io.debezium.connector.mysql;
 
 import static org.fest.assertions.Assertions.assertThat;
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 import java.util.Collections;
@@ -153,103 +152,6 @@ public class SourceInfoTest {
         assertThat(source.binlogPosition()).isEqualTo(100);
         assertThat(source.rowsToSkipUponRestart()).isEqualTo(5);
         assertThat(source.isSnapshotInEffect()).isTrue();
-    }
-
-    @Test
-    public void shouldRecoverSourceInfoFromOffsetWithFilterDataOld() {
-        final String databaseWhitelist = "a,b";
-        final String tableWhitelist = "c.foo,d.bar,d.baz";
-        Map<String, String> offset = offset(10, 10);
-        offset.put(SourceInfo.DATABASE_WHITELIST_KEY, databaseWhitelist);
-        offset.put(SourceInfo.TABLE_WHITELIST_KEY, tableWhitelist);
-
-        sourceWith(offset);
-        assertThat(source.hasFilterInfo()).isTrue();
-        assertEquals(databaseWhitelist, source.getDatabaseIncludeList());
-        assertEquals(tableWhitelist, source.getTableIncludeList());
-        // confirm other filter info is null
-        assertThat(source.getDatabaseExcludeList()).isNull();
-        assertThat(source.getTableExcludeList()).isNull();
-    }
-
-    @Test
-    public void shouldRecoverSourceInfoFromOffsetWithFilterData() {
-        final String databaseWhitelist = "a,b";
-        final String tableWhitelist = "c.foo,d.bar,d.baz";
-        Map<String, String> offset = offset(10, 10);
-        offset.put(SourceInfo.DATABASE_INCLUDE_LIST_KEY, databaseWhitelist);
-        offset.put(SourceInfo.TABLE_INCLUDE_LIST_KEY, tableWhitelist);
-
-        sourceWith(offset);
-        assertThat(source.hasFilterInfo()).isTrue();
-        assertEquals(databaseWhitelist, source.getDatabaseIncludeList());
-        assertEquals(tableWhitelist, source.getTableIncludeList());
-        // confirm other filter info is null
-        assertThat(source.getDatabaseExcludeList()).isNull();
-        assertThat(source.getTableExcludeList()).isNull();
-    }
-
-    @Test
-    public void setOffsetFilterFromFilterOld() {
-        final String databaseBlacklist = "a,b";
-        final String tableBlacklist = "c.foo, d.bar, d.baz";
-        Map<String, String> offset = offset(10, 10);
-
-        sourceWith(offset);
-        assertThat(!source.hasFilterInfo());
-
-        final Configuration configuration = Configuration.create()
-                .with(MySqlConnectorConfig.DATABASE_BLACKLIST, databaseBlacklist)
-                .with(MySqlConnectorConfig.TABLE_BLACKLIST, tableBlacklist)
-                .build();
-        source.setFilterDataFromConfig(configuration);
-
-        assertThat(source.hasFilterInfo()).isTrue();
-        assertEquals(databaseBlacklist, source.getDatabaseExcludeList());
-        assertEquals(tableBlacklist, source.getTableExcludeList());
-    }
-
-    @Test
-    public void setOffsetFilterFromFilter() {
-        final String databaseBlacklist = "a,b";
-        final String tableBlacklist = "c.foo, d.bar, d.baz";
-        Map<String, String> offset = offset(10, 10);
-
-        sourceWith(offset);
-        assertThat(!source.hasFilterInfo());
-
-        final Configuration configuration = Configuration.create()
-                .with(MySqlConnectorConfig.DATABASE_EXCLUDE_LIST, databaseBlacklist)
-                .with(MySqlConnectorConfig.TABLE_EXCLUDE_LIST, tableBlacklist)
-                .build();
-        source.setFilterDataFromConfig(configuration);
-
-        assertThat(source.hasFilterInfo()).isTrue();
-        assertEquals(databaseBlacklist, source.getDatabaseExcludeList());
-        assertEquals(tableBlacklist, source.getTableExcludeList());
-    }
-
-    @Test
-    public void shouldRecoverSourceInfoFromOffsetWithoutFilterDataIfSnapshotNewTablesIsOff() {
-        final String databaseIncludeList = "a,b";
-        final String tableIncludeList = "c.foo,d.bar,d.baz";
-        Map<String, String> offset = offset(10, 10);
-        offset.put(SourceInfo.DATABASE_INCLUDE_LIST_KEY, databaseIncludeList);
-        offset.put(SourceInfo.TABLE_INCLUDE_LIST_KEY, tableIncludeList);
-
-        sourceWith(offset);
-        assertThat(source.hasFilterInfo()).isTrue();
-
-        final Configuration configuration = Configuration.create()
-                .with(MySqlConnectorConfig.DATABASE_INCLUDE_LIST, databaseIncludeList)
-                .with(MySqlConnectorConfig.TABLE_INCLUDE_LIST, tableIncludeList)
-                .with(MySqlConnectorConfig.SNAPSHOT_NEW_TABLES, MySqlConnectorConfig.SnapshotNewTables.OFF)
-                .build();
-        source.maybeSetFilterDataFromConfig(configuration);
-
-        assertThat(source.hasFilterInfo()).isFalse();
-        assertThat(source.getDatabaseIncludeList()).isNull();
-        assertThat(source.getTableIncludeList()).isNull();
     }
 
     @Test
